Split AbilityComponent in Stateful and Stateless.
