+++ scripts/components/abilities/AbilityComponent.cs
@@ -5,6 +5,7 @@ using Incandescent.Components.Logic;
 namespace Incandescent.Components.Abilities;
 
 // TODO(calco): Refactor this to be a Node2D to allow for cool sprites and things.
+// TODO(calco): Maybe split this into a StatefulAbilityComponent and a StatelessAbilityComponent.
 /// <summary>
 /// More complex node handling the in game representation of an ability. The counterpart of <see cref="AbilityDefinition"/>.
 /// </summary>
@@ -14,6 +15,7 @@ public abstract partial class AbilityComponent : Node
 
     public abstract AbilityDefinition GetAbilityDefinition();
 
+    protected AbilityActivationData LatestActivationData { get; private set; }
     public virtual bool TryActivate(AbilityActivationData activationData)
     {
         if (!CooldownTimer.IsRunning())
@@ -23,26 +25,5 @@ public abstract partial class AbilityComponent : Node
         }
 
         return false;
-    }
-
-    protected int SelfState { get; private set; }
-    protected int FallbackState { get; private set; }
-    protected AbilityActivationData LatestActivationData { get; private set; }
-
-    public void SetStates(int selfState, int fallbackState)
-    {
-        SelfState = selfState;
-        FallbackState = fallbackState;
-    }
-
-    public virtual void Enter() { }
-    public virtual int Update()
-    {
-        return SelfState;
-    }
-    public virtual void Exit() { }
-    public virtual IEnumerator Coroutine()
-    {
-        yield break;
     }
 }
