+++ scripts/components/abilities/StatelessAbilityComponent.cs
@@ -1,0 +1,6 @@
+namespace Incandescent.Components.Abilities;
+
+public abstract partial class StatelessAbilityComponent : AbilityComponent
+{
+    public virtual void Activate() { }
+}
